Initial chinfusorExt commit
Includes the code of chinfusor_ext, a special on-request made chinfusor capable of recording passing speech requests and being controlled by shortcuts.
The crate was renamed back to sd_chinfusor, as it's no longer necessary to distinguish it from the original.
Merging these changes back to the main line is possible in the future.